fix windows and install GCC
